Features/spack test refactor (#18277)
* refactor test code into a TestSuite object and install_test module

* update mpi tests

* refactor tests suites to use content hash for name and record reproducibility info

* update unit tests and fix bugs

* Fix tests using data dir for new format
Use new `self.test_stage` object to access current data dir

Co-authored by: Tamara Dahlgren <dahlgren1@llnl.gov>

+++ lib/spack/spack/package.py
@@ -53,6 +53,7 @@ from ordereddict_backport import OrderedDict
 from spack.filesystem_view import YamlFilesystemView
 from spack.installer import \
     install_args_docstring, PackageInstaller, InstallError
+from spack.install_test import TestFailure
 from spack.util.executable import which, ProcessError
 from spack.util.prefix import Prefix
 from spack.stage import stage_prefix, Stage, ResourceStage, StageComposite
@@ -442,18 +443,6 @@ class PackageViewMixin(object):
         """
         for src, dst in merge_map.items():
             view.remove_file(src, dst)
-
-
-def test_pkg_id(spec):
-    """Build the standard install test package identifier
-
-    Args:
-        spec (Spec): instance of the spec under test
-
-    Returns:
-        (str): the install test package identifier
-    """
-    return spec.format('{name}-{version}-{hash:7}')
 
 
 def test_log_pathname(test_stage, spec):
@@ -1641,16 +1630,6 @@ class PackageBase(six.with_metaclass(PackageMeta, PackageViewMixin, object)):
 
     do_install.__doc__ += install_args_docstring
 
-    @property
-    def test_log_name(self):
-        return 'test-{0}-out.txt' \
-            .format(self.spec.format('{name}-{version}-{hash:7}'))
-
-    @property
-    def test_dir(self):
-        return self.test_stage.join(
-            self.spec.format('{name}-{version}-{hash}'))
-
     def cache_extra_test_sources(self, srcs):
         """Copy relative source paths to the corresponding install test subdir
 
@@ -1678,8 +1657,7 @@ class PackageBase(six.with_metaclass(PackageMeta, PackageViewMixin, object)):
 
     test_requires_compiler = False
     test_failures = None
-    test_log_file = None
-    test_stage = None
+    test_suite = None
 
     def do_test(self, name, remove_directory=False, dirty=False):
         if self.test_requires_compiler:
@@ -1694,25 +1672,17 @@ class PackageBase(six.with_metaclass(PackageMeta, PackageViewMixin, object)):
 
         # Clear test failures
         self.test_failures = []
-
-        self.test_stage = setup_test_stage(name)
-        self.test_log_file = test_log_pathname(self.test_stage, self.spec)
+        self.test_log_file = self.test_suite.log_file_for_spec(self.spec)
 
         def test_process():
             with tty.log.log_output(self.test_log_file) as logger:
                 with logger.force_echo():
                     tty.msg('Testing package {0}'
-                            .format(test_pkg_id(self.spec)))
+                            .format(self.test_suite.test_pkg_id(self.spec)))
 
                 # use debug print levels for log file to record commands
                 old_debug = tty.is_debug()
                 tty.set_debug(True)
-
-                # setup test directory
-                testdir = self.test_dir
-                if os.path.exists(testdir):
-                    shutil.rmtree(testdir)
-                fsys.mkdirp(testdir)
 
                 # run test methods from the package and all virtuals it provides
                 # virtuals have to be deduped by name
@@ -1730,18 +1700,23 @@ class PackageBase(six.with_metaclass(PackageMeta, PackageViewMixin, object)):
                                             for v_name in sorted(v_names)]
 
                 try:
-                    with fsys.working_dir(testdir):
+                    with fsys.working_dir(
+                            self.test_suite.test_dir_for_spec(self.spec)):
                         for spec in test_specs:
+                            self.test_suite.current_test_spec = spec
                             # Fail gracefully if a virtual has no package/tests
                             try:
                                 spec_pkg = spec.package
                             except spack.repo.UnknownPackageError:
                                 continue
 
-                            # copy test data into testdir/data
+                            # copy test data into test data dir
                             data_source = Prefix(spec_pkg.package_dir).test
-                            data_dir = os.path.join(testdir.data, spec.name)
-                            if os.path.isdir(data_source):
+                            data_dir = self.test_suite.current_test_data_dir
+                            if (os.path.isdir(data_source) and
+                                    not os.path.exists(data_dir)):
+                                # We assume data dir is used read-only
+                                # maybe enforce this later
                                 shutil.copytree(data_source, data_dir)
 
                             # grab the function for each method so we can call it
@@ -1761,8 +1736,6 @@ class PackageBase(six.with_metaclass(PackageMeta, PackageViewMixin, object)):
                     # cleanup test directory on success
                     if remove_directory:
                         shutil.rmtree(testdir)
-                        if not os.listdir(self.test_stage):
-                            shutil.rmtree(self.test_stage)
 
                 finally:
                     # reset debug level
@@ -2709,15 +2682,3 @@ class DependencyConflictError(spack.error.SpackError):
         super(DependencyConflictError, self).__init__(
             "%s conflicts with another file in the flattened directory." % (
                 conflict))
-
-
-class TestFailure(spack.error.SpackError):
-    """Raised when package tests have failed for an installation."""
-    def __init__(self, failures):
-        # Failures are all exceptions
-        msg = "%d tests failed.\n" % len(failures)
-        for failure, message in failures:
-            msg += '\n\n%s\n' % str(failure)
-            msg += '\n%s\n' % message
-
-        super(TestFailure, self).__init__(msg)
